add test modules
