+++ tests/test_modules/__init__.py
@@ -1,0 +1,7 @@
+"""
+Files in this directory contain modules that were previously implemented
+by Sensho. They were implemented separately from Roman's, but should 
+output the same results.
+Note: need to consider about initialization (how to initialize Linear and Conv
+with kernel_size=1 with the same parameters?)
+"""
